edit GoType(int64 -> int)

+++ pkg/db/model.go
@@ -40,9 +40,9 @@ type Ludoman struct {
 
 	ID              int    `pg:"ludomanId,pk"`
 	LudomanNickname string `pg:"ludomanNickname,use_zero"`
-	TgID            int64  `pg:"tgId,use_zero"`
-	Balance         int64  `pg:"balance,use_zero"`
-	Losses          int64  `pg:"losses,use_zero"`
-	TotalLost       *int64 `pg:"totalLost"`
-	TotalWon        *int64 `pg:"totalWon"`
+	TgID            int    `pg:"tgId,use_zero"`
+	Balance         int    `pg:"balance,use_zero"`
+	Losses          int    `pg:"losses,use_zero"`
+	TotalLost       *int   `pg:"totalLost"`
+	TotalWon        *int   `pg:"totalWon"`
 }
